fix routing landing
