+++ src/pages/home/index.tsx
@@ -1,0 +1,29 @@
+import Head from "next/head";
+import PersonalFinances from "../personal";
+import { useSession } from "next-auth/react";
+
+export default function Home() {
+  const { data: session } = useSession({ required: true })
+  if (session && session.user) {
+    return (
+      <>
+        <Head>
+          <title>Finanzas Personales</title>
+          <meta name="description" content="Generated by create next app" />
+          <meta name="viewport" content="width=device-width, initial-scale=1" />
+          <link rel="icon" href="/favicon.ico" />
+        </Head>
+        <main>
+          <div className="container-graphics">
+            <PersonalFinances />
+          </div>
+        </main>
+      </>
+    );
+  } else {
+    return (
+      <div>
+      </div>
+    )
+  }
+}
