+++ src/pages/index.tsx
@@ -1,29 +1,12 @@
-import Head from "next/head";
-import PersonalFinances from "./personal";
-import { useSession } from "next-auth/react";
+import React from 'react'
+import Landing from "../src-client/components/Landing/Landing"
 
-export default function Home() {
-  const { data: session } = useSession({ required: true })
-  if (session && session.user) {
-    return (
-      <>
-        <Head>
-          <title>Finanzas Personales</title>
-          <meta name="description" content="Generated by create next app" />
-          <meta name="viewport" content="width=device-width, initial-scale=1" />
-          <link rel="icon" href="/favicon.ico" />
-        </Head>
-        <main>
-          <div className="container-graphics">
-            <PersonalFinances />
-          </div>
-        </main>
-      </>
-    );
-  } else {
-    return (
-      <div>
-      </div>
-    )
-  }
+const LandingPAge = () => {
+  return (
+    <div>
+      <Landing/>
+    </div>
+  )
 }
+
+export default LandingPAge
