--- src/pages/landing.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import Landing from "../src-client/components/Landing/Landing"
-
-const LandingPAge = () => {
-  return (
-    <div>
-      <Landing/>
-    </div>
-  )
-}
-
-export default LandingPAge
